fix: issue #315 - provide endpoint policies
