+++ server/routes/admin.ts
@@ -1,4 +1,5 @@
 import { StrapiRoutes } from "../../types";
+import pluginPermissions from "../../permissions";
 
 const routes: StrapiRoutes = {
   type: 'admin',
@@ -7,26 +8,66 @@ const routes: StrapiRoutes = {
       method: 'GET',
       path: '/',
       handler: 'admin.get',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('read')],
+            },
+        }]
+      }
     },
     {
       method: 'POST',
       path: '/',
       handler: 'admin.post',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('update')],
+            },
+        }]
+      }
     },
     {
       method: 'GET',
       path: '/config',
       handler: 'admin.config',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('read')],
+            },
+        }]
+      }
     },
     {
       method: 'PUT',
       path: '/config',
       handler: 'admin.updateConfig',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('settings')],
+            },
+        }]
+      }
     },
     {
       method: 'DELETE',
       path: '/config',
       handler: 'admin.restoreConfig',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('settings')],
+            },
+        }]
+      }
     },
     {
       method: 'GET',
@@ -42,16 +83,40 @@ const routes: StrapiRoutes = {
       method: 'GET',
       path: '/:id',
       handler: 'admin.getById',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('read')],
+            },
+        }]
+      }
     },
     {
       method: 'PUT',
       path: '/:id',
       handler: 'admin.put',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('update')],
+            },
+        }]
+      }
     },
     {
       method: 'DELETE',
       path: '/:id',
       handler: 'admin.delete',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('update')],
+            },
+        }]
+      }
     },
     {
       method: 'GET',
@@ -67,24 +132,53 @@ const routes: StrapiRoutes = {
       method: 'GET',
       path: '/settings/config',
       handler: 'admin.settingsConfig',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('settings')],
+            },
+        }]
+      }
     },
     {
       method: 'GET',
       path: '/settings/restart',
       handler: 'admin.settingsRestart',
       config: {
-        policies: [],
-      },
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('settings')],
+            },
+        }]
+      }
     },
     {
       method: 'PUT',
       path: '/i18n/copy/:source/:target',
       handler: 'admin.fillFromOtherLocale',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('update')],
+            },
+        }]
+      }
     },
     {
       method: 'GET',
       path: '/i18n/item/read/:source/:target',
       handler: 'admin.readNavigationItemFromLocale',
+      config: {
+        policies: [{
+            name: "admin::hasPermissions",
+            config: {
+                actions: [pluginPermissions.render('read')],
+            },
+        }]
+      }
     },
   ]
 }
